- adds dummy http endpoint - package restructure to move Application into java/world/hello package
