+++ src/main/java/world/hello/http/controller/DummyController.java
@@ -1,0 +1,23 @@
+package world.hello.http.controller;
+
+import world.hello.http.service.DummyService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("v1/dummy")
+public class DummyController {
+
+    private final DummyService dummyService;
+
+    public DummyController(DummyService dummyService) {
+        this.dummyService = dummyService;
+    }
+
+    @GetMapping
+    public ResponseEntity<String> dummyGet() {
+        return ResponseEntity.ok(dummyService.getDummyResponse());
+    }
+}
