+++ src/main/java/world/hello/http/service/DummyService.java
@@ -1,0 +1,10 @@
+package world.hello.http.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class DummyService {
+    public String getDummyResponse() {
+        return "Hello World";
+    }
+}
