+++ src/main/java/world/hello/security/SecurityConfig.java
@@ -1,4 +1,4 @@
-package security;
+package world.hello.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
